VMM/GCM,Main,Devices/PC/DevACPI,include/VBox: Introduce facility for patching a guest in order to work around certain issues, bugref:4017 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@110627 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/ConsoleImplConfigX86.cpp
@@ -1995,7 +1995,7 @@ int Console::i_configConstructorX86(PUVM pUVM, PVM pVM, PCVMMR3VTABLE pVMM, Auto
             bool fLnxIoApicBugPatching = GetExtraDataBoth(virtualBox, pMachine,
                                                           "VBoxInternal2/LinuxIoApicPatching", &strTmp)->equals("1");
             if (fLnxIoApicBugPatching)
-                InsertConfigInteger(pCfg, "GcmPatchIdOnAcpiEnable", kGcmGstPatchId_LinuxIoApicBug);
+                InsertConfigInteger(pCfg, "GcmPatchIdOnAcpiEnable", (uint32_t)kGcmGstPatchId_LinuxIoApicBug);
 
             InsertConfigNode(pInst,    "LUN#0", &pLunL0);
             InsertConfigString(pLunL0, "Driver",               "ACPIHost");
